train.py moved to a separate file and will be copied to the pod separately
